pyproject.toml
[tool.poetry]
name = "uw-husky-directory"
version = "0.1.0"
description = "An updated version of the UW Directory"
authors = ["Thomas Thorogood <goodtom@uw.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8"
Flask = "^1.1.2"
injector = "^0.18.4"
Flask-Injector = "^0.12.3"
pydantic = {extras = ["email"], version = "^1.7.3"}
inflection = "^0.5.1"
gunicorn = "^20.0.4"
PyYAML = "^5.3.1"
requests = "^2.25.1"
devtools = "^0.6.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.1.2"
pytest-flask = "^1.1.0"
pytest-cov = "^2.10.1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
